// src/store/types.ts
export type initialStateType = {
  name: string;
  email: string;
  password: string;
  age: string;
  fullname: string;
  username: string;
  city: string;
  country: string;
  contact: string;
  jobtitle: string;
  bankname: string;
  acnumber: string;
  income: string;
  contactemail: string;
};
export type nameType = {
  name: string;
};
export type emailType = {
  email: string;
};
export type passwordType = {
  password: string;
};
export type cityType = {
  city: string;
};
export type countryType = {
  country: string;
};
export type ageType = {
  age: string;
};
export type fullnameType = {
  fullname: string;
};
export type usernameType = {
  username: string;
};
export type contactType = {
  contact: string;
};
export type jobtitleType = {
  jobtitle: string;
};
export type banknameType = {
  bankname: string;
};
export type acnumberType = {
  acnumber: string;
};
export type incomeType = {
  income: string;
};
export type contactemailType = {
  contactemail: string;
};

export type initialContentStateType = {
  price: number;
};